Fix the linux keyboard.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4951 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_Wiimote/Src/EmuMain.cpp
@@ -673,7 +673,7 @@ void ReadLinuxKeyboard()
 
 			for (int i = 0; i < LAST_CONSTANT; i++)
 			{
-				if (key == WiiMapping[g_ID].Button[i])
+				if (((int) key) == WiiMapping[g_ID].Button[i])
 					KeyStatus[i] = true;
 			}
 			break;
@@ -691,7 +691,7 @@ void ReadLinuxKeyboard()
 
 			for (int i = 0; i < LAST_CONSTANT; i++)
 			{
-				if (key == WiiMapping[g_ID].Button[i])
+				if (((int) key) == WiiMapping[g_ID].Button[i])
 					KeyStatus[i] = false;
 			}
 			break;
